+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - When importing files, the "All Files" filter is now at the top and should be selected by default.
 - Zettlr can now import and export both `textbundle` and `textpack` files. Refer to [textbundle.org](http://textbundle.org/) for more information.
 - Fixed a small bug that would render exporting of standalone files impossible.
+- Rendered Markdown links now retain the outer formatting (e.g., if you wrapped the whole link inside bold or emphasis).
 
 ## Under the Hood